refactor: authentication context
