+++ src/hooks/useAuth.ts
@@ -1,0 +1,6 @@
+import { AuthContext } from '@/context/AuthContext'
+import * as React from 'react'
+
+export function useAuth() {
+  return React.useContext(AuthContext)
+}
