+++ src/main.tsx
@@ -1,5 +1,5 @@
 import App from '@/App'
-import { AuthProvider } from '@/context/AuthContext'
+import { AuthProvider } from '@/context/AuthProvider'
 import '@/index.css'
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
